A few changes to the twist controller:
* Taking the foot off throttle while pressing on the brakes
* incrementing the brake amount by 200%
* whenever velocity higher than requested: applying brakes + taking foot off throttle

With these changes I see that the car is able to do the fake stop with less jerky stop/go behavior.

The stop/go happens because the car creeps forward automatically when there is no brake, even though there is no throttle. I will keep thinking about this and see if I can find a better solution, but for now let's see if this works fine for us.